+++ bot.py
@@ -17,7 +17,7 @@ from dotenv import load_dotenv
 if os.path.exists(".env"):
     load_dotenv()
 
-BOT_VERSION = "4.25.3"
+BOT_VERSION = "4.25.4"
 BOT_UPDATED = "2026-08-10"
 CHANGELOG_CHANNEL_ID = 1486967511839801414